+++ BetterZeeLog/BetterZeeLog.cs
@@ -14,7 +14,7 @@ using static PluginConfig;
 public sealed class BetterZeeLog : BaseUnityPlugin {
   public const string PluginGUID = "redseiko.valheim.betterzeelog";
   public const string PluginName = "BetterZeeLog";
-  public const string PluginVersion = "1.7.0";
+  public const string PluginVersion = "1.8.0";
 
   Harmony _harmony;